+++ src/bl4ckscor3/bot/bl4ckb0t/ConfigurationFile.java
@@ -90,7 +90,8 @@ public class ConfigurationFile
 			"mapsChannel=-",
 			"mapsHost=-",
 			"showPlayerCount=true",
-			"showRedditInfo=true"
+			"showRedditInfo=true",
+			"showGitHubIssueInfo=true"
 	};
 
 	/**